Clean the pawn promotion refactor

+++ include/chess.h
@@ -283,7 +283,7 @@ Bitboard	get_piece_color_control(ChessBoard *b, s8 is_black);
 void 		update_graphic_board(SDLHandle *h);
 
 /* src/pawn_promotion.c */
-s32			check_pawn_promotion(SDLHandle *handle, ChessPiece type, ChessTile tile_from, ChessTile tile_to);
+s8			check_pawn_promotion(SDLHandle *handle, ChessPiece type, ChessTile tile_from, ChessTile tile_to);
 void		do_promotion_move(SDLHandle *h, ChessTile tile_from, ChessTile tile_to, ChessPiece new_piece_type, s8 add_list);
 void		display_promotion_selection(SDLHandle *h, ChessTile tile_from, ChessTile tile_to);
 void		pawn_selection_event(SDLHandle *h);
